Added completed Paxos algorithm

+++ CurseAch/CommunicationManager.h
@@ -11,9 +11,9 @@ class CommunicationManager
 protected:
   bool networkEnabled = true;
 public:
-  virtual int broadcast(int value) = 0;
-  virtual int sendValue(int value, int to) = 0;
-  virtual Message recieveValue(int timeout) = 0;
+  virtual long long broadcast(long long value) = 0;
+  virtual long long sendValue(long long value, long long to) = 0;
+  virtual Message recieveValue(long long timeout) = 0;
   void setNetworkEnabled(bool value);
   bool getNetworkEnabled();
   CommunicationManager();
