+++ CurseAch/NetworkCommunicationManager.h
@@ -12,11 +12,11 @@ class NetworkCommunicationManager : public CommunicationManager
 private:
   HANDLE  hNamedPipe;
   std::vector<std::string> machineNames;
-  const int id = 0;
+  const long long id = 0;
 public:
-  int broadcast(int value);
-  int sendValue(int value, int to);
-  Message recieveValue(int timeout);
+  long long broadcast(long long value);
+  long long sendValue(long long value, long long to);
+  Message recieveValue(long long timeout);
   NetworkCommunicationManager();
   ~NetworkCommunicationManager();
 };
